try to fix loss calc problem #3

+++ linkfd.c
@@ -2738,7 +2738,7 @@ int lfd_linker(void)
                                 - ((int32_t) (info.channel[chan_num].local_seq_num_recv + 1)));
 //#ifdef DEBUGG
                         vtun_syslog(LOG_INFO, "loss calced seq was %"PRIu32" now %"PRIu32" loss is %"PRId16"", info.channel[chan_num].local_seq_num_recv,
-                                ntohl(local_seq_tmp), (int)info.channel[chan_num].packet_loss_counter);
+                                ntohl(local_seq_tmp), info.channel[chan_num].packet_loss_counter);
 //#endif
                     } else if (ntohl(local_seq_tmp) < info.channel[chan_num].local_seq_num_recv) {
                         vtun_syslog(LOG_INFO, "loss was %"PRId16"", info.channel[chan_num].packet_loss_counter);
